+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 ## CHANGELOG
 
-1. experimented a lot with css (flex, sticky, overflow etc).
-2. accidentally fixed the thread gaps in comments.
-3. added layout switching, press ctrl + v to switch layouts (vert split layout is still experimental).
+1. fixed bad stack display mode screen on vert display mode.